rename and move auth.dart into components/AuthWrapper.dart
